(tae): update stl
