+++ pkg/vm/engine/tae/stl/container/vector.go
@@ -1,4 +1,4 @@
-package vector
+package container
 
 import (
 	"fmt"
@@ -7,6 +7,27 @@ import (
 
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/stl"
 )
+
+func New[T any](opts ...*Options[T]) *stlVector[T] {
+	vec := &stlVector[T]{
+		buf:   make([]byte, 0),
+		slice: make([]T, 0),
+	}
+	var capacity int
+	if len(opts) > 0 {
+		opt := opts[0]
+		capacity = opt.Capacity
+		vec.alloc = opt.Allocator
+	}
+	if vec.alloc == nil {
+		vec.alloc = stl.NewSimpleAllocator[T]()
+	}
+	if capacity == 0 {
+		capacity = 4
+	}
+	vec.tryExpand(capacity)
+	return vec
+}
 
 func (vec *stlVector[T]) tryExpand(capacity int) {
 	if vec.capacity >= capacity {
