+++ pkg/vm/engine/tae/stl/container/vector_test.go
@@ -1,4 +1,4 @@
-package vector
+package container
 
 import (
 	"testing"
